Update README.md
Major update to integrate the use of a rule that was communicated to me only at the end of January - the rule of making all throws when a 4 or a 5 is rolled and storing them to choose the most strategic combination later. Indeed, beyond creating a dropdown list and subsequent choice buttons, the computer was not able to click on them and wait for updates. Now, with the addition of a few variables and timers, it can click the last created button when necessary. As for the strategy, I have my ideas, but it will require more development. Also, a minor display bug in the background has been fixed.