Update README.md
Yut Nori, also known as Yunnori, is a traditional Korean board game that is particularly popular during the Korean Lunar New Year. The game involves a board, usually made of cloth or wood, and is played with four wooden sticks, called yut sticks, which serve as dice. Each side of the sticks has a different value, dictating the movement of the players' pieces on the board.

The objective of Yut Nori is to move all of one's pieces, or "mal," around the board and back to the starting point. The game can be played by two to four players or teams, and it encourages strategic thinking and cooperation, especially in team play. The unique aspect of Yut Nori is the possibility of getting extra turns or movements based on the outcome of the stick throws, making the game unpredictable and exciting.

Yut Nori is not only a game of luck but also of strategy, as players decide the best way to move their pieces in response to the throws, and how to best utilize the special moves allowed by certain combinations of the yut sticks. It's a game that brings families and friends together, fostering a sense of community and celebration during festive occasions.